add support for SubIFD tag
This tag allows to build "TIFF trees".
Special tag which has variable length array. So we now encode that information as a 0 length array.

The libtiff convention is that you first write this tag an array of length N on a directory,
and then the next N directories written will be in this sub-directory.